Refactor Jump Kids into modules
